UI function for results boxes module

This function outputs the html tags needed to create UI for the successes, warnings, and failures results boxes.

This gives functionality to the results boxes module UI, attaching titles and populating the validation results.

results_boxes_ui(id)

results_boxes_server(input, output, session, results)

Arguments

id
The module id.

input
The input from shiny::callModule().

output
The output from shiny::callModule().

session
The session from shiny::callModule().

results
List of the validation results. If NULL, box titles will be default strings (i.e. "Successess (0)"); otherwise, the boxes will be populated with the results.

Value

The html UI for the module.

Examples

library("shiny")
library("shinydashboard")

server <- function(input, output) {
  # Create some sample results
  res <- list(
    check_pass(msg = "All good!", behavior = "Values should be >10"),
    check_fail(
      msg = "Some values are too small",
      behavior = "Values should be > 10",
      data = c(5.5, 1.3)
    )
  )
  # Show results in boxes
  callModule(results_boxes_server, "validation_results", res)
}

ui <- function(request) {
  dashboardPage(
    header = dashboardHeader(),
    sidebar = dashboardSidebar(),
    body = dashboardBody(
      includeCSS(
        system.file("app/www/custom.css", package = "dccvalidator")
      ),
      results_boxes_ui("validation_results")
    )
  )
}
if (FALSE) {
shinyApp(ui, server)
}