Check coercibility

Checks if values are coercible to a given class. Because of inconsistencies in R's built-in coercion functions (e.g. as.numeric() warns when it introduces NAs but as.logical() doesn't; as.integer() will silently remove decimal places from numeric inputs) we check only for the specific coercions we want to allow, currently: numeric, integer, or logical to string

can_coerce(values, class)

Arguments

values
Vector of values to check

class
Class of interest

Value

Boolean value; TRUE if values are coercible to class, FALSE otherwise.

Details

This function is mainly in place so that we can automatically allow numeric read lengths, pH values, etc., which are defined as strings in our annotation vocabulary but can reasonably be numbers.

This function will also return TRUE if the values are integers and the desired class is numeric.

This function will not affect validation of enumerated values, regardless of their class. It is only used by the check_type() function, which validates annotations that have a required type but no enumerated values.

See also

Examples

# Not run because function is not exported
if (FALSE) {
# Coercible:
can_coerce(1, "character")
can_coerce(TRUE, "character")
can_coerce(1L, "character")
can_coerce(1L, "numeric")

# Not coercible:
can_coerce("foo", "numeric")
can_coerce("foo", "logical")
can_coerce(2.5, "integer")
}