Check annotation values

Checks that all annotation values are valid. It does not report on values for invalid keys; see check_annotation_keys().

check_annotation_values(x, annotations, ...)

Arguments

x
An object to check.

annotations
A data frame of annotation definitions. Must contain at least three columns: key, value, and columnType.

...
Additional options to check_values()

Value

A condition object indicating whether all annotation values are valid. Invalid annotation values are included as data within the object.

Details

If the allowable annotation values are an enumerated list, check_annotation_values() compares the values in the data to the values in this list. If there is no enumerated list of values and the annotation definition merely specifies a required type, then the values are checked against that type, with values that are coercible to the correct type treated as valid (see can_coerce()).

See also

Examples

if (FALSE) {
library("synapser")
synLogin()
annots <- get_synapse_annotations()
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_values(my_file, annots)

dat <- data.frame(
  non_annotation = 5:7,
  assay = c("rnaSeq", "foo", "bar"),
  stringsAsFactors = FALSE
)
check_annotation_values(dat, annots)

fv <- synTableQuery("SELECT * FROM syn17020234")
check_annotation_values(fv, annots)

# If you don't specify an annotations data frame, these functions will
# download annotations automatically using `get_synapse_annotations()` (must
# be logged in to Synapse)
my_file <- synGet("syn11931757", downloadFile = FALSE)
check_annotation_values(my_file)

# It is possible to whitelist certain certain values, or all values for
# certain keys:
check_annotation_values(dat, whitelist_keys = "assay")
check_annotation_values(dat, whitelist_values = list(assay = c("foo")))
}